NOOP: Renames {Asset,Rune}ValueIn{Rune,Asset}WithSlip.
The new names clearly indicate the correct way to use the output of the
methods and this reveals a bug in the code at one call site. A TODO has been
added to point it out.

Also adds two more flavours of these methods (+tests) which will be used
in a follow-up change to fix the bug.

Related to: https://gitlab.com/thorchain/thornode/-/issues/1155